// src/database/schemas.ts
export interface AlertsSchema {
  /** The username or unique name of what to follow. Reddit name, twitch username, twitter name etc.. */
  id: string;
  /** The subs */
  subscriptions: AlertSub[];
}
// ...
export interface AggregatedAnalyticSchema extends AnalyticSchema {
  timestamp: number;
  guildID: string;
}

export interface AnalyticSchema {
  /** The guild id */
  id: string;
  /** The amount of messages sent on the server */
  messageCount: number;
  /** The amount of member join events */
  membersJoined: number;
  /** The amount of member leave events */
  membersLeft: number;
  /** This can be for channels, users, emojis all of the ids */
  [key: string]: number | string;
}

export interface AutoreactSchema {
  /** The channel id where this will work */
  id: string;
  /** The guild id where this is */
  guildID: string;
  /** The emoji reactions to add */
  reactions: string[];
}

export interface BlacklistedSchema {
  /** The id of the user or guild that is blacklisted. */
  id: string;
  /** Whether this is for a user or a guild. */
  type: "user" | "guild";
}

export interface ClientSchema {
  id: string;
  botID: string;
  // Bot Statistics. Using string to prevent big ints from breaking.
  messagesProcessed: string;
  messagesDeleted: string;
  messagesEdited: string;
  messagesSent: string;
  reactionsAddedProcessed: string;
  reactionsRemovedProcessed: string;
  commandsRan: string;
  feedbacksSent: string;
  automod: string;
}

export interface CommandSchema {
  /** The unique guildID-commandName */
  id: string;
  /** Whether or not the command is fully enabled */
  enabled: boolean;
  /** The exceptions to the enabled option */
  exceptionChannelIDs: string[];
  /** The exceptions to the enabled option */
  exceptionRoleIDs: string[];
  /** The guild id */
  guildID: string;
}

export interface CountingSchema {
  /** The channel id where this counting game is in. */
  id: string;
  /** The guild id where this counting game was created */
  guildID: string;
  /** The loser role id to assign when the user breaks the count */
  loserRoleID: string;
  /** Whether this game is global or only localized to the server */
  localOnly: boolean;
  /** Whether to delete non-number messages */
  deleteInvalid: boolean;
  /** The current count */
  count: number;
  /** The buffs actived for this team */
  buffs: number[];
  /** The debuffs that are on this team */
  debuffs: number[];
}

export interface EmojiSchema {
  id: string;
  /** The user id who created this emoji */
  userID: string;
  /** The id of the emoji */
  emojiID: string;
  /** The full unicode version of the emoji <:name:id> */
  fullCode: string;
  /** The guild id where the emoji is stored */
  guildID: string;
  /** The custom name for this emoji */
  name: string;
}

export interface EnterpriseSchema {
  /** The bot id */
  id: string;
  /** The server ids that this bot is allowed to have */
  guildIDs: string[];
}

export interface EventsSchema {
  /** Message id that created the event maybe. Something unique */
  id: string;
  /** The actual event id */
  eventID: number;
  /** The time the event created */
  createdAt: number;
  /** The amount of minutes to wait before starting this event. */
  minutesFromNow: number;
  /** The positions this event will allow */
  positions: {
    name: string;
    amount: number;
  }[];
  /** The guild id where this event was made */
  guildID: string;
  /** The id of the user who made this event */
  userID: string;
  /** Does the event repeat */
  isRecurring: boolean;
  /** The timestamp ms when this event is set to start */
  startsAt: number;
  /** The timestamp ms when this event is set to end */
  endsAt: number;
  /** The amount of ms to wait before the event restarts if recurring. */
  frequency: number;
  /** The title of the event */
  title: string;
  /** The description of the event */
  description: string;
  /** The platform is a place users can write custom text */
  platform: string;
  /** The game is a place users can write custom text */
  game: string;
  /** The activity is a place users can write custom text */
  activity: string;
  /** VIP can set a custom background */
  backgroundURL: string;
  /** Whether or not to remove recurring attendees. */
  removeRecurringAttendees: boolean;
  /** Whether or not to send reminders to users in dm */
  dmReminders: boolean;
  /** Whether or not to send reminders in the channel */
  channelReminders: boolean;
  /** Whether ot not to show attendees on the card */
  showAttendees: boolean;
  /** Whether or not to show the utc time on the card */
  showUTCTime: boolean;
  /** The reminders to send. */
  reminders: number[];
  /** The reminders that have already been sent */
  executedReminders: number[];
  /** Whether or not this event has started */
  hasStarted: boolean;
  /** The milliseconds of how long this event will last */
  duration: number;
  /** The max amount of users that can accept to this event */
  maxAttendees: number;
  /** The ids of the users who are attending */
  acceptedUsers: EventAttendee[];
  /** The ids of the users who are not attending */
  deniedUserIDs: string[];
  /** The ids of the users currently waiting for a spot in attending */
  waitingUsers: EventAttendee[];
  /** The ids of users who are not allowed to enter this event anymore. */
  bannedUsersIDs: string[];
  /** The role ids that are allowed to enter this event. User must have atleast 1 */
  allowedRoleIDs: string[];
  /** The role ids that are going to be mentioned for this event when it is reminded. */
  alertRoleIDs: string[];
  /** The role id that is going to be given to the user when they join this event. */
  joinRoleID: string;
  /** The ids ofthe users who want to attend but are not sure. */
  maybeUserIDs: string[];
  /** The channel id where the card is */
  cardChannelID: string;
  /** The message id where the card is */
  cardMessageID: string;
  /** The name of this template if this was changed to a template event */
  templateName: string;
}

export interface EventAttendee {
  /** The user id */
  id: string;
  /** The position the user has applied for */
  position: string;
}

export interface FeedbackSchema {
  id: string;
  userID: string;
  guildID: string;
  isBugReport: boolean;
}

export interface GachaSchema {
  ownedCharacters: GachaCharacter[];
  ownedItems: GachaItem[];
  ownedAbilities: GachaAbility[];
  gachas: number;
  foods: number[];
}

export interface GiveawaySchema {
  /** The message id will make this unique */
  id: string;
  /** The guild id where this giveaway is in */
  guildID: string;
  /** The user id who made this giveaway */
  memberID: string;
  /** The channel this giveaway is in. */
  channelID: string;
  /** The cost of Gamer coins to join. */
  costToJoin: number;
  /** The role ids that can participate */
  requiredRoleIDsToJoin: string[];
  /** The participants */
  participants: GiveawayParticipant[];
  /** THe users that have been picked, this can be winners or losers */
  pickedParticipants: GiveawayParticipant[];
  /** The user ids that have been blocked on this giveaway. */
  blockedUserIDs: string[];
  /** The time when this giveaway was created at */
  createdAt: number;
  /** The ms it would the giveaway will last */
  duration: number;
  /** The amount of users that should win this giveaway */
  amountOfWinners: number;
  /** Whether users can join multiple times */
  allowDuplicates: boolean;
  /** The time to wait before joining again. */
  duplicateCooldown: number;
  /** The emoji that is used for this giveaway */
  emoji: string;
  /** Whether or not to pick winners. If false it picks losers */
  pickWinners: boolean;
  /** The amount of time to wait in ms to pick next person */
  pickInterval: number;
  /** The notifications channel  */
  notificationsChannelID: string;
  /** The amount of ms to wait before starting this giveaway */
  delayTillStart: number;
  /** Whether or no it has started */
  hasStarted: boolean;
  /** Whether or not it has ended */
  hasEnded: boolean;
  /** Whether or not users can enter through commands */
  allowCommandEntry: boolean;
  /** Whether or not users can enter through reactions */
  allowReactionEntry: boolean;
  /** Whether or not this is a simple or advanced giveaway */
  simple: boolean;
  /** The role ids to assign when necessary */
  setRoleIDs: string[];
  /** Whether joining the giveaway requires an ingame name */
  IGN: boolean;
}

export interface GiveawayParticipant {
  memberID: string;
  joinedAt: number;
}

export interface GachaCharacter {
  id: number;
  experience: number;
  skin: number;
}

export interface GachaItem {}

export interface GachaAbility {}

export interface GuildSchema {
  // Basic settings
  id: string;
  prefix: string;
  language: string;
  tenorEnabled: boolean;

  // Leveling Settings
  xpEnabled: boolean;
  missionsDisabled: boolean;
  xpDecayDays: number;
  decayPercentange: number;
  xpPerMessage: number;
  xpPerMinuteVoice: number;
  allowedBackgroundURLs: string[];
  showMarriage: boolean;
  disabledXPChannelIDs: string[];
  disabledXPRoleIDs: string[];

  // Events Settings
  eventsAdvertiseChannelID: string;

  // Staff role ids
  adminRoleID: string;
  modRoleIDs: string[];

  // Server log channels
  logsGuildID: string;

  // Auto Embed Feature channel IDs
  autoembedChannelIDs: string[];

  // To Do Feature
  todoBacklogChannelID: string;
  todoCurrentSprintChannelID: string;
  todoNextSprintChannelID: string;
  todoArchivedChannelID: string;
  todoCompletedChannelID: string;

  // Mails feature
  mailsEnabled: boolean;
  mailsRoleIDs: string[];
  mailsGuildID: string;
  mailsLogChannelID: string;
  mailsRatingsChannelID: string;
  mailCategoryID: string;
  mailAutoResponse: string;
  mailQuestions: Question[];
  mailsSupportChannelID: string;

  // Feedback Feature
  approvalChannelID: string;
  solvedChannelID: string;
  rejectedChannelID: string;
  solvedMessage: string;
  rejectedMessage: string;
  feedbackLogChannelID: string;
  ideaChannelID: string;
  ideaQuestions: Question[];
  bugsChannelID: string;
  bugsQuestions: Question[];

  // Moderation
  publicRoleIDs: string[];
  muteRoleID: string;
  capitalPercentage: number;
  profanityEnabled: boolean;
  profanityWords: string[];
  profanityStrictWords: string[];
  profanityPhrases: string[];
  linksEnabled: boolean;
  linksChannelIDs: string[];
  linksUserIDs: string[];
  linksRoleIDs: string[];
  linksURLs: string[];
  linksRestrictedURLs: string[];

  // Verification
  verifyCategoryID: string;
  verifyEnabled: boolean;
  verifyRoleID: string;
  verifyChannelIDs: string[];
  firstMessageJSON: string;
  userAutoRoleID: string;
  botsAutoRoleID: string;
  discordVerificationStrictnessEnabled: boolean;

  // Tags feature
  disabledTagChannelIDs: string[];

  // ANalytics feature
  analyticsChannelID: string;

  // Events feature
  createEventsRoleID: string;
}

export interface ItemSchema {
  /** Unique id of this item using the message id */
  id: string;
  /** The game related to this item. */
  game: "counting";
  /** The channel id relevant to this item */
  channelID: string;
  /** The member id who bought this item. */
  memberID: string;
  /** The item id */
  itemID: number;
  /** The type of the item */
  type: "buff" | "debuff";
  /** The guild id where this item was bought. */
  guildID: string;
  /** The timestamp of when this item expires */
  expiresAt: number;
  /** For items that effect a certain amount of counts, this shows when it started */
  currentCount?: number;
}

export interface LabelSchema {
  id: string;
  /** The user id for who created the label */
  userID: string;
  /** The category id that this label is assigned to. Mails with this label will be moved to this category */
  categoryID: string;
  /** The guild id where this label is created */
  guildID: string;
  /** The main guild id for this mail system */
  mainGuildID: string;
  /** The name of the label */
  name: string;
}

export interface LevelSchema {
  /** The id is the guildID-level# */
  id: string;
  /** The guild id where it was made */
  guildID: string;
  /** The role ids to grant */
  roleIDs: string[];
}
// ...
export interface MissionSchema {
  /** UserID-commandName */
  id: string;
  userID: string;
  commandName: string;
  amount: number;
  completed: boolean;
}